requirements/staging.txt
-r base.txt

sendgrid==6.10.0
sentry-sdk==1.32.0
websockets==11.0.3
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.28.68
django-debug-toolbar==4.1.0
